ENH: Send complete model hierarchies to CLIs
Modified how model hierarchies are sent to CLIs.

Previously, a miniscene would be created that contained just the selected
hierarchy node..

Now, the miniscene contains the selected hierarchy node, its children, and its
data. In other words, the full hierarchy below the selected node is sent in the
miniscene and the models are written out with the scene. (Also, storage nodes
and display nodes are in the miniscene.)

Receiving model hiearchies from CLIs of arbitraty depth still needs to be developed.